fix: keyboard suggestions get overriden (#5)
resolves #5

The previous diff algorithm relied on the precondition that the change in text occurred
between the beginning of selection before change and the end of selection after change.

This postulate was wrong, because it didn't take into account keyboard suggestions.

The solution consists in applying a diff check to the entire line encompassed by cursor before
change.