fix: inserting a newline character in the middle of a line (#11)
resolves #11

The source of the bug is located in `DocumentDelta::getDeltasFromTextDiff` function.
If the last amongst inserted characters is a newline, retain one extra character.
This is to prevent line attributes to be shifted to next line. However, this condition
didn't take into account the possibility to break the middle of a line.

The condition is now more restrictive and exhaustive: if the last inserted character
is a newline character **and** the character right after the end of cursor before change
is a newline character, skip this character.