fix: selection discrepancy after applying line type prefix modifier (#8)
resolves #8

The solution consists in computing the next selection in `DocumentDelta::applyLineTypeToSelection`.
This selection is then send by `Document` instance to the text controller through orchestrator.

A new event has been defined, "SELECTION_OVERRIDE", which the TextBlockController is now listening to.
The latter has been refactored to gracefuly handle selection overriding.