fix: deleting a list prefix character is poorly handled (#12)
resolves #12

This ticket crossed two orthogonal issues which this patch resolves:

- `TextBlock::handleOnSelectionChange` didn't update line type with bridgeInnerInterface
- `DocumentDelta::normalize` didn't appropriately compute the number of characters to delete

From now, the later only removes the remaining characters between cursor and beginning of line.
Characters following cursor are kept to avoid a selection discrepancy. In an upcoming release,
these characters could be deleted, but a selection override must follow.