fix: text inconsistencies after normalization in TextBlockController
resolves #14

Some normalized delta updates are overriden, because of the TextInput
component triggering `onChangeText` right after applying the normalized
delta update.

To circumvent the issue, the fix consists in debouncing
`handleOnTextChanged` and `handleOnSelectionChangeEvent` by 100ms.

However, the complexity of this component is growing and an abstraction
such as streams would be beneficial to avoid state inconsistencies.

Another concern is delta updates being potentially out of sync
with selection override updates. It would be best to send explicit
instructions to the component with the expected selection after
applying delta. The component is thus in charge of dispatching the
series of update events and protecting its own state consistency.